backport cadl-ranch ARM test rename
