+++ packages/azure-http-specs/specs/azure/resource-manager/common-properties/mockapi.ts
@@ -11,7 +11,7 @@ const RESOURCE_GROUP_EXPECTED = "test-rg";
 const IDENTITY_TYPE_SYSTEM_ASSIGNED_EXPECTED = "SystemAssigned";
 const IDENTITY_TYPE_SYSTEM_USER_ASSIGNED_EXPECTED = "SystemAssigned,UserAssigned";
 const validSystemAssignedManagedIdentityResource = {
-  id: `/subscriptions/${SUBSCRIPTION_ID_EXPECTED}/resourceGroups/${RESOURCE_GROUP_EXPECTED}/providers/Azure.ResourceManager.Models.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity`,
+  id: `/subscriptions/${SUBSCRIPTION_ID_EXPECTED}/resourceGroups/${RESOURCE_GROUP_EXPECTED}/providers/Azure.ResourceManager.CommonProperties/managedIdentityTrackedResources/identity`,
   location: `${LOCATION_REGION_EXPECTED}`,
   tags: {
     tagKey1: "tagValue1",
@@ -27,7 +27,7 @@ const validSystemAssignedManagedIdentityResource = {
 };
 
 const validUserAssignedAndSystemAssignedManagedIdentityResource = {
-  id: `/subscriptions/${SUBSCRIPTION_ID_EXPECTED}/resourceGroups/${RESOURCE_GROUP_EXPECTED}/providers/Azure.ResourceManager.Models.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/identity`,
+  id: `/subscriptions/${SUBSCRIPTION_ID_EXPECTED}/resourceGroups/${RESOURCE_GROUP_EXPECTED}/providers/Azure.ResourceManager.CommonProperties/managedIdentityTrackedResources/identity`,
   location: `${LOCATION_REGION_EXPECTED}`,
   tags: {
     tagKey1: "tagValue1",
@@ -62,28 +62,27 @@ const updateExpectedIdentity = {
 };
 
 // managed identity tracked resource
-Scenarios.Azure_ResourceManager_Models_CommonTypes_ManagedIdentity_ManagedIdentityTrackedResources_get =
-  passOnSuccess({
-    uri: "/subscriptions/:subscriptionId/resourceGroups/:resourceGroup/providers/Azure.ResourceManager.Models.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/:managedIdentityResourceName",
-    method: "get",
-    request: {
-      params: {
-        subscriptionId: SUBSCRIPTION_ID_EXPECTED,
-        resourceGroup: RESOURCE_GROUP_EXPECTED,
-        managedIdentityResourceName: "identity",
-        "api-version": "2023-12-01-preview",
-      },
+Scenarios.Azure_ResourceManager_CommonProperties_ManagedIdentity_get = passOnSuccess({
+  uri: "/subscriptions/:subscriptionId/resourceGroups/:resourceGroup/providers/Azure.ResourceManager.CommonProperties/managedIdentityTrackedResources/:managedIdentityResourceName",
+  method: "get",
+  request: {
+    params: {
+      subscriptionId: SUBSCRIPTION_ID_EXPECTED,
+      resourceGroup: RESOURCE_GROUP_EXPECTED,
+      managedIdentityResourceName: "identity",
+      "api-version": "2023-12-01-preview",
     },
-    response: {
-      status: 200,
-      body: json(validSystemAssignedManagedIdentityResource),
-    },
-    kind: "MockApiDefinition",
-  });
+  },
+  response: {
+    status: 200,
+    body: json(validSystemAssignedManagedIdentityResource),
+  },
+  kind: "MockApiDefinition",
+});
 
-Scenarios.Azure_ResourceManager_Models_CommonTypes_ManagedIdentity_ManagedIdentityTrackedResources_createWithSystemAssigned =
+Scenarios.Azure_ResourceManager_CommonProperties_ManagedIdentity_createWithSystemAssigned =
   passOnSuccess({
-    uri: "/subscriptions/:subscriptionId/resourceGroups/:resourceGroup/providers/Azure.ResourceManager.Models.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/:managedIdentityResourceName",
+    uri: "/subscriptions/:subscriptionId/resourceGroups/:resourceGroup/providers/Azure.ResourceManager.CommonProperties/managedIdentityTrackedResources/:managedIdentityResourceName",
     method: "put",
     request: {
       body: {
@@ -103,9 +102,9 @@ Scenarios.Azure_ResourceManager_Models_CommonTypes_ManagedIdentity_ManagedIdenti
     kind: "MockApiDefinition",
   });
 
-Scenarios.Azure_ResourceManager_Models_CommonTypes_ManagedIdentity_ManagedIdentityTrackedResources_updateWithUserAssignedAndSystemAssigned =
+Scenarios.Azure_ResourceManager_CommonProperties_ManagedIdentity_updateWithUserAssignedAndSystemAssigned =
   passOnSuccess({
-    uri: "/subscriptions/:subscriptionId/resourceGroups/:resourceGroup/providers/Azure.ResourceManager.Models.CommonTypes.ManagedIdentity/managedIdentityTrackedResources/:managedIdentityResourceName",
+    uri: "/subscriptions/:subscriptionId/resourceGroups/:resourceGroup/providers/Azure.ResourceManager.CommonProperties/managedIdentityTrackedResources/:managedIdentityResourceName",
     method: "patch",
     request: {
       body: {
