+++ packages/azure-http-specs/specs/azure/resource-manager/resources/singleton.tsp
@@ -2,6 +2,7 @@ import "@typespec/http";
 import "@typespec/rest";
 import "@azure-tools/typespec-azure-core";
 import "@azure-tools/typespec-azure-resource-manager";
+import "@typespec/spector";
 
 using TypeSpec.Http;
 using TypeSpec.Rest;
@@ -10,7 +11,7 @@ using Azure.ResourceManager;
 using TypeSpec.OpenAPI;
 using Spector;
 
-namespace Azure.ResourceManager.Models.Resources;
+namespace Azure.ResourceManager.Resources;
 
 @singleton("default")
 model SingletonTrackedResource is TrackedResource<SingletonTrackedResourceProperties> {
@@ -28,19 +29,19 @@ model SingletonTrackedResourceProperties {
 }
 
 @armResourceOperations
-interface SingletonTrackedResources {
+interface Singleton {
   @scenario
   @scenarioDoc("""
     Resource GET operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources/default
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources/default
     Expected query parameter: api-version=2023-12-01-preview
     
     Expected response body:
     ```json
     {
-      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources/default",
+      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources/default",
       "name": "default",
-      "type": "Azure.ResourceManager.Models.Resources/singletonTrackedResources",
+      "type": "Azure.ResourceManager.Resources/singletonTrackedResources",
       "location": "eastus",
       "properties":{
         "description": "valid",
@@ -62,7 +63,7 @@ interface SingletonTrackedResources {
   @scenario
   @scenarioDoc("""
     Resource PUT operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources/default
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources/default
     Expected query parameter: api-version=2023-12-01-preview
     Expected request body:
     ```json
@@ -76,9 +77,9 @@ interface SingletonTrackedResources {
     Expected response body:
     ```json
     {
-      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources/default",
+      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources/default",
       "name": "default",
-      "type": "Azure.ResourceManager.Models.Resources/singletonTrackedResources",
+      "type": "Azure.ResourceManager.Resources/singletonTrackedResources",
       "location": "eastus",
       "properties": {
         "description": "valid",
@@ -100,7 +101,7 @@ interface SingletonTrackedResources {
   @scenario
   @scenarioDoc("""
     Resource PATCH operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources/default
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources/default
     Expected query parameter: api-version=2023-12-01-preview
     Expected request body:
     ```json
@@ -114,9 +115,9 @@ interface SingletonTrackedResources {
     Expected response body:
     ```json
     {
-      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources/default",
+      "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources/default",
       "name": "default",
-      "type": "Azure.ResourceManager.Models.Resources/singletonTrackedResources",
+      "type": "Azure.ResourceManager.Resources/singletonTrackedResources",
       "location": "eastus2",
       "properties":{
         "description": "valid2",
@@ -138,16 +139,16 @@ interface SingletonTrackedResources {
   @scenario
   @scenarioDoc("""
     Resource LIST by resource group operation.
-    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources
+    Expected path: /subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources
     Expected query parameter: api-version=2023-12-01-preview
     
     Expected response body:
     ```json
     {
       "value": [{
-        "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Models.Resources/singletonTrackedResources/default",
+        "id": "/subscriptions/00000000-0000-0000-0000-000000000000/resourceGroups/test-rg/providers/Azure.ResourceManager.Resources/singletonTrackedResources/default",
         "name": "default",
-        "type": "Azure.ResourceManager.Models.Resources/singletonTrackedResources",
+        "type": "Azure.ResourceManager.Resources/singletonTrackedResources",
         "location": "eastus",
         "properties":{
           "description": "valid",
